+++ src/pages/Usuarios/index.js
@@ -1,5 +1,5 @@
 import Login from './Login';
-import Cadastro from './Cadastro';
+import CadastroUsuario from './Cadastro';
 import ListagemUsuario from './ListagemUsuario/ListagemUsuario';
 
-export { Login, Cadastro, ListagemUsuario };
+export { Login, CadastroUsuario, ListagemUsuario };
